relocate working module
